third_party/blink/web_tests/fast/forms/select/select-multiple-elements-with-mouse-drag-with-options-less-than-size-expected.txt

Select multiple options with mouse drag with no. of options less than the size attribute

Also test mouse drag on select element without multiple attribute select only the option under mouse

P1
P2
P3
P4
P5
 
P1
P2
P3
P4
P5
Test for select element with multiple attribute
Dragging down
PASS document.getElementById("selectId").options[0].selected is true
PASS document.getElementById("selectId").options[1].selected is true
PASS document.getElementById("selectId").options[2].selected is true
PASS document.getElementById("selectId").options[3].selected is true
Dragging up
PASS document.getElementById("selectId").options[0].selected is true
PASS document.getElementById("selectId").options[1].selected is true
PASS document.getElementById("selectId").options[2].selected is true
PASS document.getElementById("selectId").options[3].selected is true
Dragging with addSelectionKey
PASS document.getElementById("selectId").options[0].selected is true
PASS document.getElementById("selectId").options[2].selected is true
PASS document.getElementById("selectId").options[3].selected is true
PASS document.getElementById("selectId").options[4].selected is true
Dragging with rangeSelectionKey
PASS document.getElementById("selectId").options[0].selected is true
PASS document.getElementById("selectId").options[1].selected is true
PASS document.getElementById("selectId").options[2].selected is true
PASS document.getElementById("selectId").options[3].selected is true
Test for select element without multiple attribute
Dragging down
PASS document.getElementById("nonmultiple").options[0].selected is false
PASS document.getElementById("nonmultiple").options[1].selected is false
PASS document.getElementById("nonmultiple").options[2].selected is false
PASS document.getElementById("nonmultiple").options[3].selected is true
Dragging up
PASS document.getElementById("nonmultiple").options[0].selected is true
PASS document.getElementById("nonmultiple").options[1].selected is false
PASS document.getElementById("nonmultiple").options[2].selected is false
PASS document.getElementById("nonmultiple").options[3].selected is false
Dragging with addSelectionKey
PASS document.getElementById("nonmultiple").options[0].selected is false
PASS document.getElementById("nonmultiple").options[1].selected is false
PASS document.getElementById("nonmultiple").options[2].selected is false
PASS document.getElementById("nonmultiple").options[3].selected is true
Dragging with rangeSelectionKey
PASS document.getElementById("nonmultiple").options[0].selected is true
PASS document.getElementById("nonmultiple").options[1].selected is false
PASS document.getElementById("nonmultiple").options[2].selected is false
PASS document.getElementById("nonmultiple").options[3].selected is false
PASS successfullyParsed is true

TEST COMPLETE
